update process module to handle status
