+++ src/logic/calculate.js
@@ -1,7 +1,7 @@
 import operate from './operate';
 
-const calculate = (stack = []) => {
-  const [LHS, operation, RHS, ...rest] = stack;
+const calculate = (queue = []) => {
+  const [LHS, operation, RHS, ...rest] = queue;
 
   if (!RHS) return operate(LHS, '+', '0');
   if (rest.length < 2) return operate(LHS, operation, RHS);
